test(api): retire the node-label tier tests E6 subsumes
TestPatchNodeLabelsIsAdmittedAtTierTwo and TestPatchNodeLabelsIsRefusedBelowTierTwo
pinned one route's tier from both sides, in the shape E4 left behind. #225's
TestEveryOperationIsGatedAtItsDeclaredTier now does that for all 53 gated
operations: it reads each one's operations-level badge and probes the assembled
router at every level to find the lowest that does not answer OPS001, so the
node-labels row is one of the rows it walks. Moving the route back to tier 3
fails it by name, which is exactly what the two tests were for.

TestAllowHeaderOffersNodePatchAtTierTwo is not subsumed and stays.
resourceWriteMethods is a second, hand-stated projection of the same tier and
the walk reads no Allow header, so a node PATCH called impactful there hides
the dashboard's labels editor on an API that accepts the edit — verified by
mutation, which the walk passes through. It moves to allow_test.go, where its
subject lives and where its own comment already said it belonged; nothing else
was left in node_labels_tier_test.go to keep the file for.

internal/mcp's cross-reference named one of the retired tests as the REST half
of the rule. It now names the walk.
